% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotParetoGrid.R
\name{plotParetoGrid}
\alias{plotParetoGrid}
\title{Visualisation of Pareto front and set}
\usage{
plotParetoGrid(fname = "ZDT1", xlim = c(0, 1), ylim = c(0, 1),
  n.grid = 100)
}
\arguments{
\item{fname}{name of the function considered,}

\item{xlim, ylim}{numeric vectors of length 2, giving the \code{x} and \code{y} coordinates ranges, default is \code{[0,1] x [0,1]},}

\item{n.grid}{number of divisions of the grid in each dimension.}
}
\description{
Plot the Pareto front and set for 2 variables 2 objectives test problems with evaluations on a grid.
}
\examples{
#------------------------------------------------------------
# Examples with test functions
#------------------------------------------------------------

plotParetoGrid("ZDT3", n.grid = 21)

plotParetoGrid("P1", n.grid = 21)

plotParetoGrid("MOP2", xlim = c(0, 1), ylim = c(0, 1), n.grid = 21) 
}
