% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictkms.R
\name{predict_kms}
\alias{predict_kms}
\title{Predict function for list of \code{\link[DiceKriging]{km}} models.}
\usage{
predict_kms(
  model,
  newdata,
  type,
  se.compute = TRUE,
  cov.compute = FALSE,
  light.return = FALSE,
  bias.correct = FALSE,
  checkNames = TRUE,
  ...
)
}
\arguments{
\item{model}{list of \code{\link[DiceKriging]{km}} models}

\item{newdata, type, se.compute, cov.compute, light.return, bias.correct, checkNames, ...}{see \code{\link[DiceKriging]{predict.km}}}
}
\description{
Predict function for list of \code{\link[DiceKriging]{km}} models.
}
\details{
So far only \code{light.return = TRUE} and \code{cov.compute = FALSE} handled.
}
